requirements.txt
numpy~=1.21.4
opencv-python>=4.1.0
split_folders
ntpath
os.path
math
Flask~=2.0.2
Werkzeug~=2.0.2
Pillow~=8.4.0